server/requirements.txt
bcrypt==4.2.1
boto3==1.42.38
dotenv==0.9.9
fastapi==0.115.14
langchain==1.2.5
langchain-experimental==0.4.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph-checkpoint-postgres==3.0.3
pgvector==0.4.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
psycopg2-binary==2.9.11
pydantic[email]==2.10.6
pyjwt==2.10.1
resend==2.21.0
pypdf==6.6.2
python-multipart==0.0.22
tavily==1.1.0
sqlmodel==0.0.31
uuid==1.30